add and adjust tests for ServiceProviderController and ServiceProviderModule
