+++ src/modules/service-provider/service-provider-api.module.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import {ConfigTestModule, DatabaseTestModule, MapperTestModule} from '../../../test/utils/index.js';
+import { ConfigTestModule, DatabaseTestModule, MapperTestModule } from '../../../test/utils/index.js';
 import { ProviderController } from './api/provider.controller.js';
 import { ServiceProviderApiModule } from './service-provider-api.module.js';
